pyproject.toml
[project]
name = "mcp-db-tarea1-ia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.26.0",
    "mysql-connector-python>=9.6.0",
]
